Crate sui_indexer_alt_framework_store_traits

Structs

CommitterWatermark

Represents the highest checkpoint for some pipeline that has been processed by the indexer framework. When read from the Store, this represents the inclusive upper bound checkpoint of data that has been written to the Store for a pipeline.

PrunerWatermark

A watermark that represents the bounds for the region that the pruner is allowed to prune, and the time in milliseconds the pruner must wait before it can begin pruning data.

ReaderWatermark

Represents the inclusive lower bound of available data in the Store for some pipeline.

Traits

Connection

Represents a database connection that can be used by the indexer framework to manage watermark operations, agnostic of the underlying store implementation.

Store

A storage-agnostic interface that provides database connections for both watermark management and arbitrary writes. The indexer framework accepts this Store implementation to manage watermarks operations through its associated Connection type. This store is also passed to the pipeline handlers to perform arbitrary writes to the store.

TransactionalStore

Extends the Store trait with transactional capabilities, to be used within the framework for atomic or transactional writes.

Functions

pipeline_task

Check that the pipeline name does not contain the store’s delimiter, and construct the string used for tracking a pipeline’s watermarks in the store. This is either the pipeline name itself, or {pipeline}{Store::DELIMITER}{task} if a task name is provided.