sui_indexer_alt_framework_store_traits

Function pipeline_task

pub fn pipeline_task<S: Store>(
    pipeline_name: &'static str,
    task_name: Option<&str>,
) -> Result<String>

Check that the pipeline name does not contain the store’s delimiter, and construct the string used for tracking a pipeline’s watermarks in the store. This is either the pipeline name itself, or {pipeline}{Store::DELIMITER}{task} if a task name is provided.